Move the writebarrier-related code into pkg/client/internal/writebarrier
